Fixed type of the handle used by _findfirst/_findnext.

+++ src/lfs.c
@@ -88,7 +88,7 @@
 typedef struct dir_data {
 	int  closed;
 #ifdef _WIN32
-	long hFile;
+	intptr_t hFile;
 	char pattern[MAX_PATH+1];
 #else
 	DIR *dir;
@@ -474,8 +474,8 @@ static int dir_iter (lua_State *L) {
 	dir_data *d = (dir_data *)luaL_checkudata (L, 1, DIR_METATABLE);
 	luaL_argcheck (L, d->closed == 0, 1, "closed directory");
 #ifdef _WIN32
-	if (d->hFile == 0L) { /* first entry */
-		if ((d->hFile = _findfirst (d->pattern, &c_file)) == -1L) {
+	if (d->hFile == 0) { /* first entry */
+		if ((d->hFile = _findfirst (d->pattern, &c_file)) == -1) {
 			lua_pushnil (L);
 			lua_pushstring (L, strerror (errno));
 			d->closed = 1;
